+++ ChysX/src/constraint/cloth/pin_constraint.cu
@@ -10,7 +10,7 @@
 #include <stdexcept>
 #include <string>
 
-#include "../sparse/block_csr_atomic.cuh"
+#include "../../sparse/block_csr_atomic.cuh"
 
 namespace chysx {
 namespace constraint {