+++ ChysX/src/constraint/cloth/triangle_stretch_constraint.cu
@@ -17,7 +17,7 @@
 #include <stdexcept>
 #include <string>
 
-#include "../sparse/block_csr_atomic.cuh"
+#include "../../sparse/block_csr_atomic.cuh"
 
 namespace chysx {
 namespace constraint {